common/encrypt.py
# -*- coding: utf-8 -*-
# @Time    : 2023/8/1 13:32
# @Author  : wwz18078
# @File    : encrypt.py
# @Software: PyCharm

import base64

from Crypto.Cipher import PKCS1_v1_5 as Cipher_pkcs1_v1_5
from Crypto.PublicKey import RSA
from django.conf import settings

PRIVATE_KEY = settings.PRIVATE_KEY
PUBLIC_KEY = settings.PUBLIC_KEY


def decrypt(crypt_data):
    rsa_key = RSA.importKey(PRIVATE_KEY)
    cipher = Cipher_pkcs1_v1_5.new(rsa_key)
    data = cipher.decrypt(base64.b64decode(crypt_data), None)
    return data
