requirements.txt
asgiref==3.4.1
cffi==1.15.1
cryptography==40.0.2
Django==3.2.20
djangorestframework==3.14.0
jwt==1.3.1
mysqlclient==2.1.1
pycparser==2.21
PyJWT==2.4.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.1.1
